Move main loop to daemon
